Fix part 1 contents not showing.
The issue seems to be that if the year of a date is unspecified, then it
defaults to the current year. Thus the start of the section was 2020 (the current
year), but the end was 2019 (the year specified). Since there are no times
that are after 2020 but before 2019, this means that the search for posts
within the time frame of the section turned up empty.